refactor social methods into social component
